move song types subquery to user resolver
