+++ projects/backend/src/__test__/mocks/SongTypeServiceMock.ts
@@ -5,6 +5,7 @@ import { SongType } from "../../models/SongType";
 export const SongTypeServiceMock = (): ISongTypeService => ({
 	getSongTypesForShare: async () => defaultSongTypes.map(SongType.fromObject),
 	getSongTypesForShares: async () => defaultSongTypes.map(SongType.fromObject),
+	getAggregatedSongTypesForUser: async () => defaultSongTypes.map(SongType.fromObject),
 	addSongTypeToShare: jest.fn(),
 	removeSongTypeFromShare: jest.fn()
 });
