+++ projects/backend/src/services/SongTypeService.ts
@@ -1,11 +1,13 @@
 import { IDatabaseClient } from "postgres-schema-builder";
 import { SongType } from "../models/SongType";
-import { flatten } from 'lodash';
+import { flatten, uniqBy } from 'lodash';
 import { SongTypesTable, ISongTypeDBResult } from "../database/schema/tables";
+import { IShareService } from "./ShareService";
 
 export interface ISongTypeService {
 	getSongTypesForShare(shareID: string): Promise<SongType[]>;
 	getSongTypesForShares(shareIDs: string[]): Promise<SongType[]>;
+	getAggregatedSongTypesForUser(userID: string): Promise<SongType[]>;
 
 	addSongTypeToShare(shareID: string, songType: SongType): Promise<void>;
 	removeSongTypeFromShare(shareID: string, songType: SongType): Promise<void>;
@@ -23,6 +25,7 @@ const filterNotRemoved = (row: ISongTypeDBResult) => row.date_removed === null;
 export class SongTypeService implements ISongTypeService {
 	constructor(
 		private readonly database: IDatabaseClient,
+		private readonly shareService: IShareService,
 	) { }
 
 	public async getSongTypesForShare(shareID: string) {
@@ -39,6 +42,13 @@ export class SongTypeService implements ISongTypeService {
 		return flatten(dbResults)
 			.filter(filterNotRemoved)
 			.map(SongType.fromDBResult);
+	}
+
+	public async getAggregatedSongTypesForUser(userID: string): Promise<SongType[]> {
+		const linkedLibraries = await this.shareService.getLinkedLibrariesOfUser(userID)
+		const aggregatedSongTypes = await this.getSongTypesForShares(linkedLibraries.map(linkedLibrary => linkedLibrary.id))
+
+		return uniqBy(aggregatedSongTypes, songType => `${songType.group}-${songType.name}`)
 	}
 
 	public async addSongTypeToShare(shareID: string, songType: SongType) {
