+++ projects/backend/src/services/services.ts
@@ -46,7 +46,7 @@ export const initServices = (config: IConfig, database: IDatabaseClient): IServi
 	const songService = new SongService(database);
 	const shareService = new ShareService(database);
 	const userService = new UserService(database);
-	const songTypeService = new SongTypeService(database);
+	const songTypeService = new SongTypeService(database, shareService);
 	const genreService = new GenreService(database, shareService);
 	const artistService = new ArtistService(songService, shareService);
 	const artistExtractor = new ArtistExtractor();
